+++ HISTORY.md
@@ -5,6 +5,10 @@ This file will contain current and previous changes to Crackle. Make sure to upd
 
 # In development
 
+## 2025-11-09
+ * better API doc
+ * events - only 'projectCreating' for now
+ 
 ## 2025-11-08
  * create HISTORY.md file like Snap!
  * helpers file not really needed - move to index.js